requirements.txt
scikit-learn
pandas
numpy
mlflow
dvc
notebook
python-box==6.0.2
pyYAML
ensure==1.0.2
types-PyYAML
Flask
joblib
boto3
python-dotenv

-e .